Rafactor soilDTO to have a dcentral point and repository select to get the central point.

+++ terra-server/src/main/java/com/terra/server/model/PolygonPoint.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PolygonPoint {
-    private float lon;
-    private float lat;
+    private double lon;
+    private double lat;
 }
